Move exec instructions and `VariableName` instruction to new modules
These had gotten glommed into `instruction/mod.rs`, and it seemed a good idea to move them out into their own modules.
