+++ packages/push/src/instruction/mod.rs
@@ -1,30 +1,20 @@
 use ordered_float::OrderedFloat;
 
-use crate::{
-    error::{Error, InstructionResult},
-    push_vm::{push_state::PushState, stack::StackError, HasStack},
-};
-use std::{fmt::Debug, fmt::Display, sync::Arc};
+use crate::push::instruction::instruction_error::PushInstructionError;
+use crate::{error::InstructionResult, push_vm::push_state::PushState};
 
-pub use self::{bool::BoolInstruction, float::FloatInstruction, int::IntInstruction};
-pub use self::{bool::BoolInstructionError, int::IntInstructionError};
+pub use self::int::IntInstructionError;
+use self::variable_name::VariableName;
+pub use self::{
+    bool::BoolInstruction, exec::ExecInstruction, float::FloatInstruction, int::IntInstruction,
+};
 
 mod bool;
+mod exec;
 mod float;
 pub mod instruction_error;
 mod int;
-
-/*
- * exec_if requires a boolean and two (additional) values on the exec stack.
- * If the bool is true, we remove the second of the two exec stack values,
- * and if it's false, we remove the first.
- */
-
-/*
- * exec_while requires a boolean and one additional value on the exec stack.
- * If the bool is true, then you push a copy of the "body" onto the exec, followed
- * by another copy of exec_while.
- */
+pub mod variable_name;
 
 /*
  * Instructions that are generic over stacks:
@@ -53,61 +43,11 @@ impl<S, E> Instruction<S> for Box<dyn Instruction<S, Error = E>> {
     }
 }
 
-// impl<F> Instruction for F
-// where
-//     F: Fn(dyn State) -> dyn State
-// {}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct VariableName(Arc<str>);
-
-impl From<&str> for VariableName {
-    fn from(s: &str) -> Self {
-        Self(Arc::from(s))
-    }
-}
-
-impl Display for VariableName {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-#[cfg(test)]
-mod variable_name_test {
-    use std::collections::HashMap;
-
-    use super::*;
-
-    #[test]
-    #[allow(clippy::unwrap_used)]
-    fn variable_name() {
-        let x = VariableName::from("x");
-        let x2 = VariableName::from("x");
-        assert_eq!(x, x2);
-        let y = VariableName::from("y");
-        assert_ne!(x, y);
-
-        let mut map = HashMap::new();
-        map.insert(x.clone(), 5);
-        map.insert(y.clone(), 7);
-
-        assert_eq!(map.get(&x).unwrap(), &5);
-        assert_eq!(map.get(&y).unwrap(), &7);
-        assert_eq!(map.len(), 2);
-
-        assert_eq!(map.get(&x2).unwrap(), &5);
-
-        let z = VariableName::from("z");
-        assert_eq!(map.get(&z), None);
-    }
-}
-
 #[derive(Clone, Eq, PartialEq)]
 #[non_exhaustive]
 pub enum PushInstruction {
     InputVar(VariableName),
-    Block(Vec<PushInstruction>),
+    Exec(ExecInstruction),
     BoolInstruction(BoolInstruction),
     IntInstruction(IntInstruction),
     FloatInstruction(FloatInstruction),
@@ -152,7 +92,7 @@ impl Instruction<PushState> for PushInstruction {
                 //   Or add a `with_input` that returns the new state and keep `push_input`?
                 state.with_input(var_name)
             }
-            Self::Block(block) => block.perform(state),
+            Self::Exec(_) => todo!(),
             Self::BoolInstruction(i) => i.perform(state),
             Self::IntInstruction(i) => i.perform(state),
             Self::FloatInstruction(i) => i.perform(state),
