+++ packages/push/src/instruction/variable_name.rs
@@ -1,0 +1,47 @@
+use std::fmt::Display;
+use std::sync::Arc;
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub struct VariableName(Arc<str>);
+
+impl From<&str> for VariableName {
+    fn from(s: &str) -> Self {
+        Self(Arc::from(s))
+    }
+}
+
+impl Display for VariableName {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+#[cfg(test)]
+pub(crate) mod variable_name_test {
+    use std::collections::HashMap;
+
+    use super::*;
+
+    #[test]
+    #[allow(clippy::unwrap_used)]
+    fn variable_name() {
+        let x = VariableName::from("x");
+        let x2 = VariableName::from("x");
+        assert_eq!(x, x2);
+        let y = VariableName::from("y");
+        assert_ne!(x, y);
+
+        let mut map = HashMap::new();
+        map.insert(x.clone(), 5);
+        map.insert(y.clone(), 7);
+
+        assert_eq!(map.get(&x).unwrap(), &5);
+        assert_eq!(map.get(&y).unwrap(), &7);
+        assert_eq!(map.len(), 2);
+
+        assert_eq!(map.get(&x2).unwrap(), &5);
+
+        let z = VariableName::from("z");
+        assert_eq!(map.get(&z), None);
+    }
+}
